Try to fix attribute error
The listener was randomlyl throwing "'Condition' object has no attribute
'_lock'" exceptions inside the threading library code, which is very
weird. I added a lock to the listener to try to avoid creating multiples
therads in a race condition, which seems to have fixed the problem or
pushed ir somewhere else.