Removed administrator check for macs.
I tested it. It isn't needed. Enforcing administrator is actually worse for
people that have alternate keyboard layouts based on my testing. When you
enforce administrator, I think the library is assuming the administrator's
keyboard layout which is probably qwerty and I don't know how to or if it's
possible to change that to something else. But when you run the library under
the actual user, it uses that user's keyboard mapping.